+++ socket.lisp
@@ -1,6 +1,6 @@
 (in-package #:com.aragaer.pa-brain)
 
-(defun create-socket-stream (socket-path)
+(defun create-unix-socket-stream (socket-path)
   #+rawsock
   (rawsock:open-unix-socket-stream socket-path :direction :io)
   #+sbcl
@@ -9,3 +9,19 @@
     (sb-bsd-sockets:socket-make-stream socket :input t :output t :buffering :line))
   #-(or rawsock sbcl)
   (error "not implemented"))
+
+(defun create-tcp-socket-stream (host port)
+  #+sbcl
+  (let ((socket (make-instance 'sb-bsd-sockets:inet-socket :type :stream :protocol :tcp))
+        (addr (sb-bsd-sockets:make-inet-address host)))
+    (sb-bsd-sockets:socket-connect socket addr port)
+    (sb-bsd-sockets:socket-make-stream socket :input t :output t :buffering :line))
+  #-sbcl
+  (error "not implemented"))
+
+(defun create-socket-stream (addr)
+  (if-let ((colon (position #\: addr)))
+      (let ((host (subseq addr 0 colon))
+            (port (parse-integer (subseq addr (+ 1 colon)))))
+        (create-tcp-socket-stream host port))
+    (create-unix-socket-stream addr)))